# specs/001-event-album-media/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Event Album Media Sharing API
  version: 0.1.0
paths:
  /api/events:
    post:
      summary: Create an event and its primary album
      operationId: createEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
      responses:
        '201':
          description: Event and album created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/events/{eventId}/share:
    post:
      summary: Generate or regenerate QR/link share access
      operationId: createShareAccess
      parameters:
        - $ref: '#/components/parameters/EventId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShareAccessRequest'
      responses:
        '201':
          description: Share access created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareAccess'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/share/{shareSlug}/join:
    post:
      summary: Join an album through QR/link and receive a nickname
      operationId: joinShareAccess
      parameters:
        - name: shareSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Guest session and nickname ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestJoin'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/upload-intents:
    post:
      summary: Create a signed upload intent for album media
      operationId: createUploadIntent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUploadIntentRequest'
      responses:
        '201':
          description: Upload intent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadIntent'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/albums/{albumId}/media:
    get:
      summary: List album media visible to the requester
      operationId: listAlbumMedia
      parameters:
        - $ref: '#/components/parameters/AlbumId'
        - name: tag
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Media list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/MediaItem'
    post:
      summary: Confirm uploaded media and add it to the album
      operationId: confirmMediaUpload
      parameters:
        - $ref: '#/components/parameters/AlbumId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmMediaUploadRequest'
      responses:
        '201':
          description: Media item created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaItem'
  /api/media/{mediaId}/tags:
    post:
      summary: Add a text or people tag to media
      operationId: addMediaTag
      parameters:
        - $ref: '#/components/parameters/MediaId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTagRequest'
      responses:
        '201':
          description: Tag created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
  /api/tags/{tagId}:
    delete:
      summary: Remove a tag if requester is creator or organizer
      operationId: removeTag
      parameters:
        - name: tagId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Tag removed
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  parameters:
    EventId:
      name: eventId
      in: path
      required: true
      schema:
        type: string
    AlbumId:
      name: albumId
      in: path
      required: true
      schema:
        type: string
    MediaId:
      name: mediaId
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Request validation failed
    Unauthorized:
      description: Authentication required
    Forbidden:
      description: Requester is not allowed to perform this action
    NotFound:
      description: Resource not found or share access unavailable
  schemas:
    CreateEventRequest:
      type: object
      required: [name, eventType, startsAt, albumTitle]
      properties:
        name:
          type: string
        eventType:
          type: string
          enum: [wedding, birthday, conference, concert, reunion, party, custom]
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        albumTitle:
          type: string
        description:
          type: string
        location:
          type: string
        visibilityEndsAt:
          type: string
          format: date-time
    EventCreated:
      type: object
      required: [eventId, albumId, share]
      properties:
        eventId:
          type: string
        albumId:
          type: string
        share:
          $ref: '#/components/schemas/ShareAccess'
    CreateShareAccessRequest:
      type: object
      required: [permissions]
      properties:
        permissions:
          type: array
          items:
            type: string
            enum: [view, upload, tag]
        expiresAt:
          type: string
          format: date-time
    ShareAccess:
      type: object
      required: [id, slug, link, qrCodeUrl, state, permissions]
      properties:
        id:
          type: string
        slug:
          type: string
        link:
          type: string
        qrCodeUrl:
          type: string
        state:
          type: string
          enum: [active, disabled, expired]
        permissions:
          type: array
          items:
            type: string
    GuestJoin:
      type: object
      required: [eventId, albumId, participantId, nickname]
      properties:
        eventId:
          type: string
        albumId:
          type: string
        participantId:
          type: string
        nickname:
          type: string
    CreateUploadIntentRequest:
      type: object
      required: [albumId, filename, mediaType, byteSize]
      properties:
        albumId:
          type: string
        filename:
          type: string
        mediaType:
          type: string
          enum: [image, video]
        byteSize:
          type: integer
    UploadIntent:
      type: object
      required: [uploadId, objectKey, uploadUrl]
      properties:
        uploadId:
          type: string
        objectKey:
          type: string
        uploadUrl:
          type: string
    ConfirmMediaUploadRequest:
      type: object
      required: [uploadId, objectKey, mediaType]
      properties:
        uploadId:
          type: string
        objectKey:
          type: string
        mediaType:
          type: string
          enum: [image, video]
    MediaItem:
      type: object
      required: [id, albumId, mediaType, status, nickname, uploadedAt]
      properties:
        id:
          type: string
        albumId:
          type: string
        mediaType:
          type: string
        status:
          type: string
        nickname:
          type: string
        uploadedAt:
          type: string
          format: date-time
        thumbnailUrl:
          type: string
    CreateTagRequest:
      type: object
      required: [tagType, displayName]
      properties:
        tagType:
          type: string
          enum: [text, person]
        displayName:
          type: string
    Tag:
      type: object
      required: [id, mediaId, tagType, displayName, creatorNickname]
      properties:
        id:
          type: string
        mediaId:
          type: string
        tagType:
          type: string
        displayName:
          type: string
        creatorNickname:
          type: string